+++ src/StreamFilters/AdviceFilter.php
@@ -281,15 +281,17 @@ class AdviceFilter extends AbstractFilter
 
                             $pointcutExpression->setJoinpoint($joinpoint);
 
-                            // "straighten out" structure and function referenced by the pointcut to avoid regex within generated code
-                            $pointcutExpression->getPointcut()->straightenExpression($functionDefinition);
+                            // "straighten out" structure and function referenced by the pointcut to avoid regex within generated code.
+                            // Same here with the cloning: we do not want to influence the matching process with working on references
+                            $expressionPointcut = clone $pointcutExpression->getPointcut();
+                            $expressionPointcut->straightenExpression($functionDefinition);
 
                             // add the weaving pointcut into the expression
                             $pointcutExpression->setPointcut(new AndPointcut(
                                 AdvisePointcut::TYPE . str_replace('\\\\', '\\', '(\\' . $advice->getQualifiedName() . ')'),
-                                $pointcutExpression->getPointcut()->getType() . '(' . $pointcutExpression->getPointcut()->getExpression() . ')'
+                                $expressionPointcut->getType() . '(' . $expressionPointcut->getExpression() . ')'
                             ));
-                            $pointcutExpression->setString($pointcutExpression->getPointcut()->getExpression());
+                            $pointcutExpression->setString($expressionPointcut->getExpression());
 
                             // add it to our result list
                             $pointcutExpressions->add($pointcutExpression);
@@ -323,7 +325,7 @@ class AdviceFilter extends AbstractFilter
         $code = ReservedKeywords::METHOD_INVOCATION_OBJECT . ' = new \AppserverIo\Doppelgaenger\Entities\MethodInvocation(
             ';
 
-        // add the original method call to the callback chain so it can be integrated, add it and get add the context
+        // add the original method call to the callback chain so it can be integrated, add it and add the context
         if ($functionDefinition->isStatic()) {
             $contextCode = '__CLASS__';
 
@@ -393,7 +395,7 @@ class AdviceFilter extends AbstractFilter
     /**
      * Will generate and advice chain of callbacks to the given around pointcut expressions
      *
-     * @param array              $sortedPointcutExpressions Pointcut expressions sorted by their joinpoint's code hooks
+     * @param array              $sortedPointcutExpressions Pointcut expressions sorted by their join-point's code hooks
      * @param FunctionDefinition $functionDefinition        Definition of the function to inject invocation code into
      *
      * @return array
@@ -405,7 +407,7 @@ class AdviceFilter extends AbstractFilter
         $callbackChain = array();
         if (isset($sortedPointcutExpressions[Around::ANNOTATION])) {
             foreach ($sortedPointcutExpressions[Around::ANNOTATION] as $aroundExpression) {
-                $callbackChain = array_merge($callbackChain, $aroundExpression->getPointcut()->getCallbackChain());
+                $callbackChain = array_merge($callbackChain, $aroundExpression->getPointcut()->getCallbackChain($functionDefinition));
             }
         }
 
@@ -429,7 +431,7 @@ class AdviceFilter extends AbstractFilter
      */
     protected function sortPointcutExpressions($pointcutExpressions)
     {
-        // sort by joinpoint code hooks
+        // sort by join-point code hooks
         $sortedPointcutExpressions = array();
         foreach ($pointcutExpressions as $pointcutExpression) {
             $sortedPointcutExpressions[$pointcutExpression->getJoinpoint()->getCodeHook()][] = $pointcutExpression;